requirements.txt
torch>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.8
openai>=1.0
python-dotenv>=1.0
